requirements.txt
apache-airflow==2.1.4
psycopg2~=2.9.6
python-dotenv~=1.0.0
requests~=2.31.0
